// cgocore-ui/src/app/features/models/bookings/views/full/DatedFlightBookingFullView.ts
import { DatedFlightBookingForm } from 'src/app/features/models/bookings/forms/DatedFlightBookingForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightBookingFullView {

id: number;
form: DatedFlightBookingForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}