// cgocore-ui/src/app/features/models/schedule/views/full/DatedFlightFullView.ts
import { DatedFlightForm } from 'src/app/features/models/schedule/forms/DatedFlightForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightFullView {

id: number;
form: DatedFlightForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}